+++ test/cmdline/cmdline_test.cpp
@@ -75,16 +75,16 @@ static std::string toString(const char (&value)[N]) {
 
 TEST_F(CmdlineTest, marker1) {
     // line marker of syntax error
-    const char *msg = R"((string):1: [syntax error] expected `=', `:'
-var a
-      ^
-)";
+    const char *msg = "(string):4: [syntax error] expected `=', `:'\n   \n   ^\n";
     ASSERT_NO_FATAL_FAILURE(this->expect(ds("-c", "var a   \n    \\\n   \t  \t  \n   "), 1, "", msg));
+
+    msg = "(string):1: [syntax error] expected `=', `:'\nvar a    \n         ^\n";
+    ASSERT_NO_FATAL_FAILURE(this->expect(ds("-c", "var a    "), 1, "", msg));
 
     auto result = ds("-c", "{").execAndGetResult(false);
     ASSERT_EQ(1, result.status.value);
     ASSERT_EQ("", result.out);
-    ASSERT_STREQ("{\n  ^\n", strchr(result.err.c_str(), '\n') + 1);
+    ASSERT_STREQ("{\n ^\n", strchr(result.err.c_str(), '\n') + 1);
 
     result = ds("-c", "\n);").execAndGetResult(false);
     ASSERT_EQ(1, result.status.value);